requirements.txt
requests==2.31.0
python-telegram-bot==20.7
pytz==2023.3
